requirements.txt
Django==5.2.11
cloudinary==1.41.0
django-modeltranslation==0.19.14
django-htmx==1.27.0
django-solo==2.5.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
whitenoise==6.8.2
Pillow==12.1.0
